Moved JS and CSS files to a component folder
