+++ src/App.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Headline } from "./components/Headline";
-import Chart from "./components/chart";
-import LatestNews from "./components/LatestNews";
-import KeyStats from "./components/KeyStats";
-import { Company } from "./components/Company";
-import { Footer } from "./components/Footer";
+import { Headline } from "./components/headline";
+import { Chart } from "./components/chart";
+import { LatestNews } from "./components/latestnews";
+import { KeyStats } from "./components/keystats";
+import { Company } from "./components/company";
+import { Footer } from "./components/footer";
 
 const App = () => {
   const selectedStock = useSelector(state => state.stockData.selectedStock);
