+++ src/components/company/index.js
@@ -1,0 +1,1 @@
+export { Company } from "./Company";
