+++ src/components/errormessage/index.js
@@ -1,0 +1,1 @@
+export { ErrorMessage } from "./ErrorMessage";
