+++ src/components/peers/index.js
@@ -1,0 +1,1 @@
+export { Peers } from "./Peers";
