+++ src/components/search/index.js
@@ -1,0 +1,1 @@
+export { Search } from "./Search";
