+++ src/components/stockticker/StockTicker.css
