+++ src/styles/Global.css
@@ -32,18 +32,6 @@ h1.title {
   padding: 0 0 3px;
 }
 
-.loading-spinner {
-  border: 10px solid #ffffff31;
-  border-radius: 50%;
-  border-top: 10px solid #95a4d1;
-  width: 50px;
-  height: 50px;
-  -webkit-animation: spin 2s linear infinite;
-  animation: spin 2s linear infinite;
-  margin-left: auto;
-  margin-right: auto;
-}
-
 @keyframes spin {
   0% {
     transform: rotate(0deg);
